+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/ManagerCompendium/YouthTeam.cs
@@ -21,7 +21,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.ManagerCompendium
         /// <summary>
         /// Gets or sets the ID.
         /// </summary>
-        public uint YouthTeamId { get; set; }
+        public long YouthTeamId { get; set; }
 
         /// <summary>
         /// Gets or sets the name.
