+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/TeamDetails/Region.cs
@@ -16,7 +16,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the Region ID.
         /// </summary>
-        public uint RegionId { get; set; }
+        public long RegionId { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the Region.
