+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/TeamDetails/SupporterTeam.cs
@@ -16,12 +16,12 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the League Id.
         /// </summary>
-        public uint LeagueId { get; set; }
+        public long LeagueId { get; set; }
 
         /// <summary>
         /// Gets or sets the LeagueLevelUnit Id.
         /// </summary>
-        public uint LeagueLevelUnitId { get; set; }
+        public long LeagueLevelUnitId { get; set; }
 
         /// <summary>
         /// Gets or sets the LeagueLevelUnit Name.
@@ -41,7 +41,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the Team Id.
         /// </summary>
-        public uint TeamId { get; set; }
+        public long TeamId { get; set; }
 
         /// <summary>
         /// Gets or sets the Team Name.
@@ -51,7 +51,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the User Id.
         /// </summary>
-        public uint UserId { get; set; }
+        public long UserId { get; set; }
 
         #endregion Public Properties
     }
