+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/WorldDetails/Country.cs
@@ -23,7 +23,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.WorldDetails
         /// <summary>
         /// Gets or sets the Country ID.
         /// </summary>
-        public uint CountryId { get; set; }
+        public long CountryId { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the Country.
